[refactor] Split loader from context
Improve separation of concerns
- locales/loader is responsible for managing the known locales
- locales/context is responsible for managing the current locale
