+++ src/lib/create/from-anything.js
@@ -7,7 +7,7 @@ import isDate from '../utils/is-date';
 import map from '../utils/map';
 import { createInvalid } from './valid';
 import { Moment, isMoment } from '../moment/constructor';
-import { getLocale } from '../locale/locales';
+import { getLocale } from '../locale/context';
 import { hooks } from '../utils/hooks';
 import checkOverflow from './check-overflow';
 import { isValid } from './valid';
