+++ src/lib/locale/lists.js
@@ -1,5 +1,5 @@
 import isNumber from '../utils/is-number';
-import { getLocale } from './locales';
+import { getLocale } from './context';
 import { createUTC } from '../create/utc';
 
 function get (format, index, field, setter) {
